🔭 Stellar: Corrected camera databases and added Hesiodus Ray
- Audited QHY and ZWO camera databases, removing non-existent 'ghost' mono models for color-only sensors (IMX071, IMX094, IMX128, IMX410, IMX482, IMX485, IMX662, IMX664, IMX676, IMX678, IMX715).
- Corrected peak Quantum Efficiency (QE) for color models (QHY268C, QHY600C, ASI2600MC, ASI6200MC) to 80% per manufacturer documentation.
- Added 'Hesiodus Ray' lunar feature tracking in `apts/skyfield_searches/lunar.py` at colongitude 18.0.
- Verified all changes with unit tests and addressed code review feedback regarding orphaned factory methods and popular model convenience.
