+++ apts/opticalequipment/camera/vendors/zwo.py
@@ -1476,40 +1476,6 @@ class ZwoCamera(Camera):
             "type": "type_camera",
             "width": 1920,
         },
-        "ZWO_ASI_482MM": {
-            "bf_role": "end",
-            "brand": "ZWO",
-            "cside_gender": "",
-            "cside_thread": "",
-            "mass": 150,
-            "name": "ASI482MM",
-            "optical_length": 6.5,
-            "reversible": False,
-            "tside_gender": "Female",
-            "tside_thread": "M42",
-            "type": "type_camera",
-        },
-        "ZWO_ASI_482MM_V2": {
-            "bf_role": "end",
-            "brand": "ZWO",
-            "cside_gender": "",
-            "cside_thread": "",
-            "full_well_e": 51500,
-            "height": 1080,
-            "mass": 155,
-            "name": "ASI482MM V2",
-            "optical_length": 6.5,
-            "pixel_size_um": 5.8,
-            "quantum_efficiency_pct": 85,
-            "read_noise_e": 1.5,
-            "reversible": False,
-            "sensor_height_mm": 6.26,
-            "sensor_width_mm": 11.13,
-            "tside_gender": "Female",
-            "tside_thread": "M42",
-            "type": "type_camera",
-            "width": 1920,
-        },
         "ZWO_ASI_485MC": {
             "bf_role": "end",
             "brand": "ZWO",
@@ -1526,27 +1492,6 @@ class ZwoCamera(Camera):
             "reversible": False,
             "sensor_height_mm": 6.26,
             "sensor_width_mm": 11.13,
-            "tside_gender": "Female",
-            "tside_thread": "M42",
-            "type": "type_camera",
-            "width": 3840,
-        },
-        "ZWO_ASI_678MM_V2": {
-            "bf_role": "end",
-            "brand": "ZWO",
-            "cside_gender": "",
-            "cside_thread": "",
-            "full_well_e": 11270,
-            "height": 2160,
-            "mass": 155,
-            "name": "ASI678MM V2",
-            "optical_length": 6.5,
-            "pixel_size_um": 2.0,
-            "quantum_efficiency_pct": 83,
-            "read_noise_e": 0.6,
-            "reversible": False,
-            "sensor_height_mm": 4.32,
-            "sensor_width_mm": 7.68,
             "tside_gender": "Female",
             "tside_thread": "M42",
             "type": "type_camera",
@@ -1572,19 +1517,6 @@ class ZwoCamera(Camera):
             "tside_thread": "M42",
             "type": "type_camera",
             "width": 3840,
-        },
-        "ZWO_ASI_485MM": {
-            "bf_role": "end",
-            "brand": "ZWO",
-            "cside_gender": "",
-            "cside_thread": "",
-            "mass": 150,
-            "name": "ASI485MM",
-            "optical_length": 6.5,
-            "reversible": False,
-            "tside_gender": "Female",
-            "tside_thread": "M42",
-            "type": "type_camera",
         },
         "ZWO_ASI_533MC": {
             "bf_role": "end",
@@ -2254,27 +2186,6 @@ class ZwoCamera(Camera):
             "type": "type_camera",
             "width": 3840,
         },
-        "ZWO_ASI_678MM": {
-            "bf_role": "end",
-            "brand": "ZWO",
-            "cside_gender": "",
-            "cside_thread": "",
-            "full_well_e": 11270,
-            "height": 2160,
-            "mass": 150,
-            "name": "ASI678MM",
-            "optical_length": 6.5,
-            "pixel_size_um": 2.0,
-            "quantum_efficiency_pct": 83,
-            "read_noise_e": 0.6,
-            "reversible": False,
-            "sensor_height_mm": 4.32,
-            "sensor_width_mm": 7.68,
-            "tside_gender": "Female",
-            "tside_thread": "M42",
-            "type": "type_camera",
-            "width": 3840,
-        },
         "ZWO_ASI_715MC": {
             "bf_role": "end",
             "brand": "ZWO",
@@ -2305,40 +2216,6 @@ class ZwoCamera(Camera):
             "height": 2192,
             "mass": 155,
             "name": "ASI715MC V2",
-            "optical_length": 6.5,
-            "pixel_size_um": 1.45,
-            "quantum_efficiency_pct": 80,
-            "read_noise_e": 0.72,
-            "reversible": False,
-            "sensor_height_mm": 3.18,
-            "sensor_width_mm": 5.6,
-            "tside_gender": "Female",
-            "tside_thread": "M42",
-            "type": "type_camera",
-            "width": 3864,
-        },
-        "ZWO_ASI_715MM": {
-            "bf_role": "end",
-            "brand": "ZWO",
-            "cside_gender": "",
-            "cside_thread": "",
-            "mass": 150,
-            "name": "ASI715MM",
-            "optical_length": 6.5,
-            "reversible": False,
-            "tside_gender": "Female",
-            "tside_thread": "M42",
-            "type": "type_camera",
-        },
-        "ZWO_ASI_715MM_V2": {
-            "bf_role": "end",
-            "brand": "ZWO",
-            "cside_gender": "",
-            "cside_thread": "",
-            "full_well_e": 6030,
-            "height": 2192,
-            "mass": 155,
-            "name": "ASI715MM V2",
             "optical_length": 6.5,
             "pixel_size_um": 1.45,
             "quantum_efficiency_pct": 80,
@@ -2634,9 +2511,6 @@ class ZwoCamera(Camera):
     def ZWO_ASI_678MC(cls):
         return cls.from_database(cls._DATABASE["ZWO_ASI_678MC"])
 
-    @classmethod
-    def ZWO_ASI_678MM(cls):
-        return cls.from_database(cls._DATABASE["ZWO_ASI_678MM"])
 
     @classmethod
     def ZWO_ASI_662MC(cls):
@@ -2658,17 +2532,6 @@ class ZwoCamera(Camera):
     def ZWO_ASI_482MC(cls):
         return cls.from_database(cls._DATABASE["ZWO_ASI_482MC"])
 
-    @classmethod
-    def ZWO_ASI_482MM(cls):
-        return cls.from_database(cls._DATABASE["ZWO_ASI_482MM"])
-
-    @classmethod
-    def ZWO_ASI_485MC(cls):
-        return cls.from_database(cls._DATABASE["ZWO_ASI_485MC"])
-
-    @classmethod
-    def ZWO_ASI_485MM(cls):
-        return cls.from_database(cls._DATABASE["ZWO_ASI_485MM"])
 
     @classmethod
     def ZWO_ASI_432MM(cls):
@@ -2686,9 +2549,6 @@ class ZwoCamera(Camera):
     def ZWO_ASI_715MC(cls):
         return cls.from_database(cls._DATABASE["ZWO_ASI_715MC"])
 
-    @classmethod
-    def ZWO_ASI_715MM(cls):
-        return cls.from_database(cls._DATABASE["ZWO_ASI_715MM"])
 
     @classmethod
     def ZWO_ASI_676MC(cls):
@@ -2805,6 +2665,7 @@ class ZwoCamera(Camera):
     @classmethod
     def ZWO_ASI_071MC_Pro_6_bolt_mount(cls):
         return cls.from_database(cls._DATABASE["ZWO_ASI_071MC_Pro_6_bolt_mount"])
+
 
     @classmethod
     def ZWO_ASI_2400MC_Pro_6_bolt_mount(cls):
@@ -2958,9 +2819,6 @@ class ZwoCamera(Camera):
     def ZWO_ASI_715MC_V2(cls):
         return cls.from_database(cls._DATABASE["ZWO_ASI_715MC_V2"])
 
-    @classmethod
-    def ZWO_ASI_715MM_V2(cls):
-        return cls.from_database(cls._DATABASE["ZWO_ASI_715MM_V2"])
 
     @classmethod
     def ZWO_ASI_585MC_V2(cls):
@@ -2982,17 +2840,10 @@ class ZwoCamera(Camera):
     def ZWO_ASI_678MC_V2(cls):
         return cls.from_database(cls._DATABASE["ZWO_ASI_678MC_V2"])
 
-    @classmethod
-    def ZWO_ASI_678MM_V2(cls):
-        return cls.from_database(cls._DATABASE["ZWO_ASI_678MM_V2"])
 
     @classmethod
-    def ZWO_ASI_482MC_V2(cls):
-        return cls.from_database(cls._DATABASE["ZWO_ASI_482MC_V2"])
-
-    @classmethod
-    def ZWO_ASI_482MM_V2(cls):
-        return cls.from_database(cls._DATABASE["ZWO_ASI_482MM_V2"])
+    def ZWO_ASI_485MC(cls):
+        return cls.from_database(cls._DATABASE["ZWO_ASI_485MC"])
 
     @classmethod
     def ZWO_ASI_585MC_Pro(cls):
