+++ apts/place.py
@@ -2,28 +2,28 @@ import datetime
 import logging
 import math
 from functools import lru_cache
-from importlib import resources
 from math import copysign as copysign
 from math import radians as rad
 from typing import Any, Iterable, Optional, Tuple, cast
 
-import matplotlib.font_manager as font_manager
 import numpy as np
 import pandas as pd
-import pytz
 from dateutil import tz
 from skyfield import almanac
 from skyfield.api import Topos
 from timezonefinder import TimezoneFinder
 
 from apts.cache import get_ephemeris, get_timescale
-from apts.config import get_dark_mode
-from apts.constants.graphconstants import get_plot_style
 from apts.constants.twilight import Twilight
-from apts.i18n import gettext_
 from apts.light_pollution import LightPollution
-from apts.utils.plot import Utils as PlotUtils
 
+from .skyfield_searches.visibility import (
+    get_twilight_time_utc,
+    next_rising_time_utc,
+    next_setting_time_utc,
+    previous_rising_time_utc,
+    previous_setting_time_utc,
+)
 from .utils.planetary import (
     get_moon_age,
     get_moon_distance,
@@ -37,133 +37,11 @@ from .weather import Weather
 logger = logging.getLogger(__name__)
 
 
-@lru_cache(maxsize=1024)
-def _get_twilight_time_utc(lat, lon, elevation, start_date, twilight, event):
-    ts = get_timescale()
-    eph = get_ephemeris()
-    location = Topos(latitude_degrees=lat, longitude_degrees=lon, elevation_m=float(elevation))
-    t0 = ts.utc(start_date)
-    t1 = ts.utc(start_date + datetime.timedelta(days=2))
-
-    f = almanac.dark_twilight_day(eph, location)
-    times, events = almanac.find_discrete(t0, t1, f)
-
-    # Define transitions for evening (set) and morning (rise)
-    if event == "set":  # Evening: getting darker
-        transitions = {
-            Twilight.CIVIL: (3, 2),  # Civil -> Nautical
-            Twilight.NAUTICAL: (2, 1),  # Nautical -> Astronomical
-            Twilight.ASTRONOMICAL: (1, 0),  # Astronomical -> Night
-        }
-    else:  # Morning: getting lighter
-        transitions = {
-            Twilight.ASTRONOMICAL: (0, 1),  # Night -> Astronomical
-            Twilight.NAUTICAL: (1, 2),  # Astronomical -> Nautical
-            Twilight.CIVIL: (2, 3),  # Nautical -> Civil
-        }
-
-    trans = transitions.get(twilight)
-    if trans is None:
-        return None
-    prev_event, next_event = trans
-
-    # The first event is the state at t0
-    previous_y = f(t0)
-    if times is not None and events is not None:
-        for t, y in zip(times, events):
-            if previous_y == prev_event and y == next_event:
-                return t.utc_datetime().replace(tzinfo=pytz.UTC)
-            previous_y = y
-
-    return None
-
-
-@lru_cache(maxsize=1024)
-def _previous_setting_time_utc(lat, lon, elevation, obj_name, start, horizon_degrees=None):
-    ts = get_timescale()
-    eph = get_ephemeris()
-    location = Topos(latitude_degrees=lat, longitude_degrees=lon, elevation_m=float(elevation))
-    obj = eph[obj_name]
-    t0 = ts.utc(start - datetime.timedelta(days=2))
-    t1 = ts.utc(start)
-    if horizon_degrees is not None:
-        f = almanac.risings_and_settings(eph, obj, location, horizon_degrees=horizon_degrees)
-    else:
-        f = almanac.risings_and_settings(eph, obj, location)
-    t, y = almanac.find_discrete(t0, t1, f)
-
-    if t is None:
-        return None
-
-    settings = [ti for ti, yi in zip(t, y) if yi == 0]
-    if settings:
-        return settings[-1].utc_datetime().replace(tzinfo=pytz.UTC)
-    return None
-
-
-@lru_cache(maxsize=1024)
-def _next_rising_time_utc(lat, lon, elevation, obj_name, start, horizon_degrees=None):
-    ts = get_timescale()
-    eph = get_ephemeris()
-    location = Topos(latitude_degrees=lat, longitude_degrees=lon, elevation_m=float(elevation))
-    obj = eph[obj_name]
-    t0 = ts.utc(start)
-    t1 = ts.utc(start + datetime.timedelta(days=2))
-    if horizon_degrees is not None:
-        f = almanac.risings_and_settings(eph, obj, location, horizon_degrees=horizon_degrees)
-    else:
-        f = almanac.risings_and_settings(eph, obj, location)
-    t, y = almanac.find_discrete(t0, t1, f)
-
-    if t is not None:
-        for ti, yi in zip(t, y):
-            if yi == 1:  # Rising
-                return ti.utc_datetime().replace(tzinfo=pytz.UTC)
-    return None
-
-
-@lru_cache(maxsize=1024)
-def _next_setting_time_utc(lat, lon, elevation, obj_name, start, horizon_degrees=None):
-    ts = get_timescale()
-    eph = get_ephemeris()
-    location = Topos(latitude_degrees=lat, longitude_degrees=lon, elevation_m=float(elevation))
-    obj = eph[obj_name]
-    t0 = ts.utc(start)
-    t1 = ts.utc(start + datetime.timedelta(days=2))
-    if horizon_degrees is not None:
-        f = almanac.risings_and_settings(eph, obj, location, horizon_degrees=horizon_degrees)
-    else:
-        f = almanac.risings_and_settings(eph, obj, location)
-    t, y = almanac.find_discrete(t0, t1, f)
-
-    if t is not None:
-        for ti, yi in zip(t, y):
-            if yi == 0:  # Setting
-                return ti.utc_datetime().replace(tzinfo=pytz.UTC)
-    return None
-
-
-@lru_cache(maxsize=1024)
-def _previous_rising_time_utc(lat, lon, elevation, obj_name, start, horizon_degrees=None):
-    ts = get_timescale()
-    eph = get_ephemeris()
-    location = Topos(latitude_degrees=lat, longitude_degrees=lon, elevation_m=float(elevation))
-    obj = eph[obj_name]
-    t0 = ts.utc(start - datetime.timedelta(days=2))
-    t1 = ts.utc(start)
-    if horizon_degrees is not None:
-        f = almanac.risings_and_settings(eph, obj, location, horizon_degrees=horizon_degrees)
-    else:
-        f = almanac.risings_and_settings(eph, obj, location)
-    t, y = almanac.find_discrete(t0, t1, f)
-
-    if t is None:
-        return None
-
-    risings = [ti for ti, yi in zip(t, y) if yi == 1]
-    if risings:
-        return risings[-1].utc_datetime().replace(tzinfo=pytz.UTC)
-    return None
+get_twilight_time_utc = lru_cache(maxsize=1024)(get_twilight_time_utc)
+previous_setting_time_utc = lru_cache(maxsize=1024)(previous_setting_time_utc)
+next_rising_time_utc = lru_cache(maxsize=1024)(next_rising_time_utc)
+next_setting_time_utc = lru_cache(maxsize=1024)(next_setting_time_utc)
+previous_rising_time_utc = lru_cache(maxsize=1024)(previous_rising_time_utc)
 
 
 class TFProxy:
@@ -177,9 +55,6 @@ class TFProxy:
 
 
 class Place:
-    MOON_FONT = font_manager.FontProperties(
-        fname=str(resources.files("apts").joinpath("data/moon_phases.ttf")), size=50
-    )
     TF = TFProxy()
 
     def __init__(
@@ -495,7 +370,7 @@ class Place:
         self._add_extra_weather_info()
 
     def _previous_setting_time(self, obj_name, start, horizon_degrees=None):
-        res = _previous_setting_time_utc(
+        res = previous_setting_time_utc(
             self.lat_decimal,
             self.lon_decimal,
             self.elevation,
@@ -508,7 +383,7 @@ class Place:
         return None
 
     def _next_setting_time(self, obj_name, start, horizon_degrees=None):
-        res = _next_setting_time_utc(
+        res = next_setting_time_utc(
             self.lat_decimal,
             self.lon_decimal,
             self.elevation,
@@ -521,7 +396,7 @@ class Place:
         return None
 
     def _previous_rising_time(self, obj_name, start, horizon_degrees=None):
-        res = _previous_rising_time_utc(
+        res = previous_rising_time_utc(
             self.lat_decimal,
             self.lon_decimal,
             self.elevation,
@@ -534,7 +409,7 @@ class Place:
         return None
 
     def _next_rising_time(self, obj_name, start, horizon_degrees=None):
-        res = _next_rising_time_utc(
+        res = next_rising_time_utc(
             self.lat_decimal,
             self.lon_decimal,
             self.elevation,
@@ -561,7 +436,7 @@ class Place:
             return self.date.utc_datetime()  # type: ignore
 
     def _get_twilight_time(self, start_date, twilight: Twilight, event: str):
-        res = _get_twilight_time_utc(
+        res = get_twilight_time_utc(
             self.lat_decimal,
             self.lon_decimal,
             self.elevation,
@@ -757,100 +632,9 @@ class Place:
         return df
 
     def plot_sun_path(self, dark_mode_override: Optional[bool] = None, **args):
-        if dark_mode_override is not None:
-            effective_dark_mode = dark_mode_override
-        else:
-            effective_dark_mode = get_dark_mode()
+        from .plotting.path import generate_plot_sun_path
 
-        style = get_plot_style(effective_dark_mode)
-
-        def add_marker(
-            ax, label, position, text_color, grid_color
-        ):  # Pass ax and colors
-            ax.axvline(position, color=grid_color, linestyle="--", linewidth=1)
-            ax.text(
-                position,
-                1,
-                label,
-                weight="bold",
-                horizontalalignment="center",
-                color=text_color,
-            )
-
-        data = self.sun_path()
-        if self.lat_decimal < 0:
-            data["Azimuth"] = (data["Azimuth"] + 180) % 360 - 180
-
-        passed_ax = args.pop("ax", None)  # Renamed to avoid confusion
-
-        plot_kwargs = {
-            "x": "Azimuth",
-            "y": "Sun altitude",
-            "title": "Sun altitude",  # This title is styled later by direct ax.set_title
-            "style": ".-",
-            **args,  # Pass through any other user-supplied keyword arguments
-        }
-
-        if passed_ax is not None:
-            plot_kwargs["ax"] = passed_ax  # Add 'ax' to kwargs only if it was provided
-            data.plot(**plot_kwargs)  # Plot on the provided ax
-            ax = passed_ax  # Use the axes that was passed in
-            fig = ax.figure  # Get figure from provided ax
-        else:
-            ax = data.plot(**plot_kwargs)  # Let pandas create a new ax and figure
-            fig = ax.figure  # Get figure from newly created ax
-
-        fig.patch.set_facecolor(style["FIGURE_FACE_COLOR"])
-        ax.set_facecolor(style["AXES_FACE_COLOR"])
-
-        if ax.lines:  # Style the main moon path line
-            ax.lines[0].set_color(style["TEXT_COLOR"])
-
-        ax.set_title(
-            gettext_("Sun Path on {date}").format(
-                date=self.date.utc_datetime().strftime("%Y-%m-%d")  # type: ignore
-            ),
-            color=style["TEXT_COLOR"],
-        )
-
-        # Add cardinal direction and set x-axis limits based on hemisphere
-        if self.lat_decimal < 0:  # Southern Hemisphere
-            add_marker(ax, "W", -90, style["TEXT_COLOR"], style["GRID_COLOR"])
-            add_marker(ax, "N", 0, style["TEXT_COLOR"], style["GRID_COLOR"])
-            add_marker(ax, "E", 90, style["TEXT_COLOR"], style["GRID_COLOR"])
-            ax.set_xlim(135, -135)  # Inverted for South-up view, SE to SW
-        else:  # Northern Hemisphere
-            add_marker(ax, "E", 90, style["TEXT_COLOR"], style["GRID_COLOR"])
-            add_marker(ax, "S", 180, style["TEXT_COLOR"], style["GRID_COLOR"])
-            add_marker(ax, "W", 270, style["TEXT_COLOR"], style["GRID_COLOR"])
-            ax.set_xlim(45, 315)
-
-        PlotUtils.annotate_plot(
-            ax,
-            gettext_("Altitude [°]"),
-            effective_dark_mode,
-            self.local_timezone,
-            x_label=gettext_("Azimuth [°]"),
-        )
-
-        # Plot horizon
-        ax.axhspan(0, -50, color=style["GRID_COLOR"], alpha=0.3)
-        ax.locator_params(nbins=20)
-        ax.set_ylim(bottom=-10, top=90)
-
-        # Plot time for altitudes
-        for _, obj_row in data.dropna().iloc[::6, :].iterrows():
-            altitude = obj_row["Sun altitude"]
-            azimuth = obj_row["Azimuth"]
-            local_time = obj_row["Local_time"]
-            if altitude > 0:
-                ax.annotate(
-                    local_time,
-                    (azimuth + copysign(10, azimuth - 180) + 10, altitude + 1),
-                    color=style["TEXT_COLOR"],
-                )
-        PlotUtils.style_legend(ax, style)
-        return ax
+        return generate_plot_sun_path(self, dark_mode_override, **args)
 
     def __getstate__(self):
         state = self.__dict__.copy()
@@ -878,152 +662,6 @@ class Place:
         self.moon = self.eph["moon"]
 
     def plot_moon_path(self, dark_mode_override: Optional[bool] = None, **args):
-        if dark_mode_override is not None:
-            effective_dark_mode = dark_mode_override
-        else:
-            effective_dark_mode = get_dark_mode()
+        from .plotting.path import generate_plot_moon_path
 
-        style = get_plot_style(effective_dark_mode)
-
-        def add_marker(
-            ax, label, position, text_color, grid_color
-        ):  # Pass ax and colors
-            ax.axvline(position, color=grid_color, linestyle="--", linewidth=1)
-            ax.text(
-                position,
-                1,
-                label,
-                weight="bold",
-                horizontalalignment="center",
-                color=text_color,
-            )
-
-        data = self.moon_path()
-        if self.lat_decimal < 0:
-            data["Azimuth"] = (data["Azimuth"] + 180) % 360 - 180
-
-        passed_ax = args.pop("ax", None)  # Renamed to avoid confusion
-
-        plot_kwargs = {
-            "x": "Azimuth",
-            "y": "Moon altitude",
-            "title": "Moon altitude",  # This title is styled later by direct ax.set_title
-            "style": ".-",
-            **args,  # Pass through any other user-supplied keyword arguments
-        }
-
-        if passed_ax is not None:
-            plot_kwargs["ax"] = passed_ax  # Add 'ax' to kwargs only if it was provided
-            data.plot(**plot_kwargs)  # Plot on the provided ax
-            ax = passed_ax  # Use the axes that was passed in
-            fig = ax.figure  # Get figure from provided ax
-        else:
-            ax = data.plot(**plot_kwargs)  # Let pandas create a new ax and figure
-            fig = ax.figure  # Get figure from newly created ax
-
-        fig.patch.set_facecolor(style["FIGURE_FACE_COLOR"])
-        ax.set_facecolor(style["AXES_FACE_COLOR"])
-
-        if ax.lines:  # Style the main moon path line
-            ax.lines[0].set_color(style["TEXT_COLOR"])
-
-        # Add cardinal direction and set x-axis limits based on hemisphere
-        if self.lat_decimal < 0:  # Southern Hemisphere
-            add_marker(ax, "W", -90, style["TEXT_COLOR"], style["GRID_COLOR"])
-            add_marker(ax, "N", 0, style["TEXT_COLOR"], style["GRID_COLOR"])
-            add_marker(ax, "E", 90, style["TEXT_COLOR"], style["GRID_COLOR"])
-            ax.set_xlim(135, -135)  # Inverted for South-up view
-        else:  # Northern Hemisphere
-            add_marker(ax, "E", 90, style["TEXT_COLOR"], style["GRID_COLOR"])
-            add_marker(ax, "S", 180, style["TEXT_COLOR"], style["GRID_COLOR"])
-            add_marker(ax, "W", 270, style["TEXT_COLOR"], style["GRID_COLOR"])
-            ax.set_xlim(45, 315)
-        PlotUtils.annotate_plot(
-            ax,
-            gettext_("Altitude [°]"),
-            effective_dark_mode,
-            self.local_timezone,
-            x_label=gettext_("Azimuth [°]"),
-        )
-
-        # Plot horizon
-        ax.axhspan(0, -50, color=style["GRID_COLOR"], alpha=0.3)
-        ax.locator_params(nbins=20)
-        ax.set_ylim(bottom=-10, top=90)
-
-        # --- Plot Moon Phase Icon and Illumination Text ---
-        # Position icon differently based on hemisphere to avoid overlap with path
-        if self.lat_decimal < 0:  # Southern Hemisphere, transit is North
-            icon_x_position = 0
-            icon_y_position = 20
-            text_y_position = icon_y_position - 13
-        else:  # Northern Hemisphere, transit is South
-            icon_x_position = 180
-            icon_y_position = 10
-            text_y_position = -3
-
-        if effective_dark_mode:
-            # In dark mode, draw a solid circle first, then the shadow on top.
-            ax.plot(
-                icon_x_position,
-                icon_y_position,
-                marker="o",
-                markersize=45,
-                color=style["TEXT_COLOR"],
-                linestyle="None",
-            )
-            ax.text(
-                icon_x_position,
-                icon_y_position,
-                self._moon_phase_letter(),
-                fontproperties=Place.MOON_FONT,
-                horizontalalignment="center",
-                verticalalignment="center",
-                color=style["AXES_FACE_COLOR"],
-            )
-        else:
-            # In light mode, just draw the phase character.
-            ax.text(
-                icon_x_position,
-                icon_y_position,
-                self._moon_phase_letter(),
-                fontproperties=Place.MOON_FONT,
-                horizontalalignment="center",
-                verticalalignment="center",
-                color=style["TEXT_COLOR"],
-            )
-        ax.text(
-            icon_x_position,
-            text_y_position,
-            f"{self.moon_illumination():.0f}%",
-            color=style["TEXT_COLOR"],
-            alpha=0.7,
-            horizontalalignment="center",
-        )
-
-        # Plot time for altitudes, adjusting for hemisphere
-        for _, obj_row in data.dropna().iloc[::6, :].iterrows():
-            altitude = obj_row["Moon altitude"]
-            azimuth = obj_row["Azimuth"]
-            local_time = obj_row["Local_time"]
-
-            if altitude > 0:
-                if self.lat_decimal < 0:  # Southern Hemisphere
-                    # Adjust offset direction to avoid text crossing the plot edges
-                    offset_direction = azimuth
-                    x_offset = copysign(10, offset_direction) + 10
-                else:  # Northern Hemisphere
-                    x_offset = copysign(10, azimuth - 180) - 8
-
-                ax.annotate(
-                    local_time,
-                    (azimuth + x_offset, altitude + 1),
-                    color=style["TEXT_COLOR"],
-                )
-        ax.set_title(
-            gettext_("Moon Path")
-            + f" on {self.date.utc_datetime().strftime('%Y-%m-%d')}",  # type: ignore
-            color=style["TEXT_COLOR"],
-        )
-        PlotUtils.style_legend(ax, style)
-        return ax
+        return generate_plot_moon_path(self, dark_mode_override, **args)
