+++ tests/plotting/test_plot_path.py
@@ -3,17 +3,17 @@ from unittest.mock import MagicMock, patch
 from apts.plot import plot_sun_and_moon_path
 
 
-def test_plot_sun_and_moon_path_sun():
+@patch("apts.plotting.path.generate_plot_sun_path")
+def test_plot_sun_and_moon_path_sun(mock_generate_plot_sun_path):
     mock_observation = MagicMock()
     mock_observation.sun_observation = True
-    with patch.object(mock_observation.place, "plot_sun_path") as mock_plot_sun_path:
-        plot_sun_and_moon_path(mock_observation)
-        mock_plot_sun_path.assert_called_once()
+    plot_sun_and_moon_path(mock_observation)
+    mock_generate_plot_sun_path.assert_called_once_with(mock_observation.place, None)
 
 
-def test_plot_sun_and_moon_path_moon():
+@patch("apts.plotting.path.generate_plot_moon_path")
+def test_plot_sun_and_moon_path_moon(mock_generate_plot_moon_path):
     mock_observation = MagicMock()
     mock_observation.sun_observation = False
-    with patch.object(mock_observation.place, "plot_moon_path") as mock_plot_moon_path:
-        plot_sun_and_moon_path(mock_observation)
-        mock_plot_moon_path.assert_called_once()
+    plot_sun_and_moon_path(mock_observation)
+    mock_generate_plot_moon_path.assert_called_once_with(mock_observation.place, None)
